requirements.txt
# Core scientific libraries
openpyxl
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.13.1
kneed==0.8.5

# Dimensionality reduction
umap-learn==0.5.6

# Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0

# Optional (for MDS performance boost)
joblib==1.4.2

# For interactive dashboard (bonus)
streamlit==1.35.0

kneed==0.8.5
